\name{aggrisk-package}
\alias{aggrisk-package}
\alias{aggrisk}
\docType{package}
\title{
Estimate individual-level risk using individual case data and spatially aggregated control data.
}
\description{
This package estimates regression parameters in the intensity function of an inhomogeneous
spatial Poisson process, in situations where information on risk-factors for the cases is available at the
individual level, whereas control data is only available at a spatially aggregated level. Estimation is performed using 
weighted estimating equations, using data-driven methods to select the weights. A formal test is available to to detect non-Poisson
behaviour in the underlying point process.
}
\details{
\tabular{ll}{
Package: \tab aggrisk\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-09\cr
License: \tab GPL (version 2 or later)\cr
LazyLoad: \tab yes\cr
}

}
\author{Michelle Stanton \email{m.stanton@lancaster.ac.uk}, Yongtao Guan \email{yongtao.guan@yale.edu} }
\references{Diggle, P.J., Guan, Y., Hart, C.A., Paize, F. 
and Stanton, M.C. (2010) Estimating 
individual-level risk in spatial epidemiology using spatially aggregated information on the population at risk. \emph{Journal of the American Statistical Association}, \bold{105} 1394--1402. \cr \cr
Stanton, M.C., Taylor-Robinson, D., Harris, D., Paize, F., Makwana, N., Hackett, S.J., Baines, P.B., Riordan, F.A.I., Marzouk, O., Thomson, A.P.J., Diggle, P., Hart, C.A and Carrol, E.A. (2011) Meningococcal disease in children in Merseyside, England: a 31 year descriptive study. \emph{Submitted} }
\keyword{ package }


